+++ apiyizazun/src/main/java/com/amp/apiyizazun/repositories/FoodRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.amp.apiyizazun.models.Food;
 
-public interface FoodRepository extends CrudRepository<Food, Long> {
+public interface FoodRepository extends CrudRepository<Food, Integer> {
 
 }
